refactor: PinInput, FormatInput의 로직을 hook으로 관심사 분리
