+++ src/shared/components/Input/hooks/index.ts
@@ -1,3 +1,6 @@
 export * from './useInputFieldsValues';
 export * from './useInputRefs';
 export * from './useInputValues';
+export * from './useFormatInputField';
+export * from './useFormatInputTextCounter';
+export * from './usePinInputField';
